+++ demisto_sdk/commands/create_artifacts/tests/content_artifacts_creator_test.py
@@ -134,8 +134,8 @@ def mock_git(mocker):
     from demisto_sdk.commands.common.content import Content
 
     # Mock git working directory
-    mocker.patch.object(Content, "git")
-    Content.git().working_tree_dir = TEST_CONTENT_REPO
+    mocker.patch.object(Content, "git_util")
+    Content.git_util().repo.working_tree_dir = TEST_CONTENT_REPO
     yield
 
 
@@ -407,8 +407,10 @@ def mock_single_pack_git(mocker):
     from demisto_sdk.commands.common.content import Content
 
     # Mock git working directory
-    mocker.patch.object(Content, "git")
-    Content.git().working_tree_dir = TEST_DATA / "content_repo_with_alternative_fields"
+    mocker.patch.object(Content, "git_util")
+    Content.git_util().repo.working_tree_dir = (
+        TEST_DATA / "content_repo_with_alternative_fields"
+    )
     yield